+++ packages/next/src/client/components/app-router.tsx
@@ -52,7 +52,7 @@ import { hasBasePath } from '../has-base-path'
 import { getSelectedParams } from './router-reducer/compute-changed-path'
 import type { FlightRouterState } from '../../server/app-render/types'
 import { useNavFailureHandler } from './nav-failure-handler'
-import type { AppRouterActionQueue } from '../../shared/lib/router/action-queue'
+import type { AppRouterActionQueue } from './app-router-instance'
 import { prefetch as prefetchWithSegmentCache } from './segment-cache'
 import { getRedirectTypeFromError, getURLFromRedirectError } from './redirect'
 import { isRedirectError, RedirectType } from './redirect-error'
